refactor: rename ESP32-DevKit-V4 to ESP32-DevKitC-V4
