+++ .github/workflows/esp32_build_examples.yml
@@ -18,12 +18,12 @@ jobs:
       - name: Checkout repository
         uses: actions/checkout@v4.1.7
 
-      - name: Build ESP32-DevKit-V4 examples
+      - name: Build ESP32-DevKitC-V4 examples
         run: |
             . /opt/esp/idf/export.sh
-            idf.py -C examples/esp32/ESP32-DevKit-V4/hello_world/ build
-            idf.py -C examples/esp32/ESP32-DevKit-V4/identify_chip/ build
-            idf.py -C examples/esp32/ESP32-DevKit-V4/fw_update/ build
+            idf.py -C examples/esp32/ESP32-DevKitC-V4/hello_world/ build
+            idf.py -C examples/esp32/ESP32-DevKitC-V4/identify_chip/ build
+            idf.py -C examples/esp32/ESP32-DevKitC-V4/fw_update/ build
 
       - name: Build ESP32-S3-DevKitC-1 examples
         run: |
