+++ examples/esp32/ESP32-DevKitC-V4/fw_update/sdkconfig.defaults
