+++ examples/esp32/ESP32-DevKitC-V4/hello_world/sdkconfig.defaults
