+++ examples/esp32/ESP32-DevKitC-V4/identify_chip/components/libtropic/CMakeLists.txt
