+++ tests/functional/esp32/ESP32-DevKitC-V4/CMakeLists.txt
@@ -8,7 +8,7 @@ set(CMAKE_TOOLCHAIN_FILE $ENV{IDF_PATH}/tools/cmake/toolchain-${TARGET}.cmake)
 # Project configuration for this board
 project(libtropic_functional_tests_esp32_devkit_v4
 	VERSION 0.1.0
-	DESCRIPTION "Functional tests in ESP-IDF environment using ESP32-DevKit-V4."
+	DESCRIPTION "Functional tests in ESP-IDF environment using ESP32-DevKitC-V4."
 	LANGUAGES C)
 
 include(${CMAKE_CURRENT_LIST_DIR}/../common.cmake)
