+++ tests/functional/esp32/ESP32-DevKitC-V4/main.c
@@ -1,6 +1,6 @@
 /**
  * @file main.c
- * @brief Common entrypoint for running functional tests against ESP32-DevKitC V4.
+ * @brief Common entrypoint for running functional tests against ESP32-DevKitC-V4.
  * @copyright Copyright (c) 2020-2025 Tropic Square s.r.o.
  *
  * @license For the license see LICENSE.md in the root directory of this source tree.
